+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 # Next
 
+## [0.6.0](https://github.com/DenTelezhkin/Ariadne/releases/tag/0.6.0)
+
 ### Added 
 
 * `ReplaceNavigationTransition` to encapsulate `UINavigationController.setViewControllers(_:animated:)` method call.